robot_graphSLAMNavigation_andobstacleavoidance
This project includes a demonstration of graph SLAM algorithm which is used to navigate the current position of the robot. The code also includes a function for obstacle avoidance and the number of obstacles and their positions can be added by the user. The world is a 2d world with size 100 cross 100. The included code can be opened using google colab.